Add files
Modified app.py -> app2.py to show content of the .json file giving docid; however, it didn't work yet. The return_content.py show a workable example that if given a valid docid (i.e. 100), it should return the content of the corresponding file. So, I think the reason it's not working with the search engine is that the get_item() does not have at all/ not valid docid passed into it as parameter. But I don't know how to solve it yet.